+++ app/controllers/host_factories_controller.rb
@@ -12,6 +12,8 @@ class HostFactoriesController < ApplicationController
   # This requires the host factory's token in the Authorization header.
   def create_host
     raise(ArgumentError, "Invalid resource kind: #{hf_token.resource.kind}") unless hf_token.resource.kind == 'host_factory'
+    params.require(:id)
+    raise(ArgumentError, "Invalid id: #{params[:id]}") if params[:id].start_with?('conjur', '/conjur')
 
     host, api_key = do_create_host
     response = host.as_json
@@ -27,7 +29,7 @@ class HostFactoriesController < ApplicationController
     role = host_factory.role
     HostBuilder.new(
       host_factory.account,
-      params.require(:id),
+      params[:id],
       role,
       role.layers,
       params.except(:id)